modernize exception-printing test (#8300)
